Move test-files around for better namings
