--- resources/maps/routing.fmi
@@ -1,41 +1,0 @@
-# node_count
-8
-
-# edge_count
-16
-
-# nodes
-#       id     osm-id        lat       lon
-         0       -     0_1234567 0_1234567 # a
-         1       -     1_2345678 1_2345678 # b
-         2       -     2_3456789 2_3456789 # c
-         3       -     3_4567890 3_4567890 # d
-         4       -     4_5678901 4_5678901 # e
-         5       -     5_6789012 5_6789012 # f
-         6       -     6_7890123 6_7890123 # g
-         7       -     7_8901234 7_8901234 # h
-
-# edges
-#   src-id     dst-id distance ? maxspeed
-         1          0        1 _       30 # b -> a
-         1          2        1 _       30 # b -> c
-
-         2          0        1 _       30 # c -> a
-         2          1        1 _       30 # c -> b
-
-         3          1        1 _       30 # d -> b
-         3          4        2 _       30 # d -> e
-         3          7        1 _       30 # d -> h
-
-         4          3        2 _       30 # e -> d
-         4          5        1 _       30 # e -> f
-
-         5          4        1 _       30 # f -> e
-         5          7        1 _       30 # f -> h
-
-         6          4        1 _       30 # g -> e
-         6          5        1 _       30 # g -> f
-
-         7          2        4 _       30 # h -> c
-         7          3        1 _       30 # h -> d
-         7          5        1 _       30 # h -> f
