+++ tests/main.rs
@@ -1,1 +1,3 @@
+mod network;
 mod parsing;
+mod routing;
