+++ tests/network/graph_construction/pbf.rs
@@ -1,0 +1,5 @@
+#[test]
+#[ignore]
+fn isle_of_man() {
+    // TODO check graph structure
+}
