+++ tests/parsing/pbf.rs
@@ -5,7 +5,7 @@ use log::error;
 use osmgraphing::Parser;
 
 #[test]
-fn parse() {
+fn isle_of_man() {
     let path = OsString::from("resources/maps/isle-of-man_2019-09-05.osm.pbf");
     let _graph = match Parser::parse(&path) {
         Ok(graph) => graph,
@@ -14,10 +14,4 @@ fn parse() {
             return;
         }
     };
-}
-
-#[test]
-#[ignore]
-fn graph_construction() {
-    // TODO check graph structure
 }
