--- tests/parsing/pbf/mod.rs
@@ -1,1 +1,0 @@
-mod isle_of_man;
